devblog-service/requirements.txt
Flask
Flask-Migrate
flask-restx
Flask-SQLAlchemy
Werkzeug==2.1.2
python-decouple==3.4
